requirements.txt
python-dotenv==0.19.2
requests==2.27.1
qrcode==7.3.1
django-qrcode==0.3
opencv-python==4.5.5.64
pyTelegramBotAPI==4.4.0
telebot==0.0.4